Separate getInitialMapPosition to it's own function
