+++ src/modules/search/components/result/map/MapResults.tsx
@@ -10,6 +10,7 @@ import { calculateApartmentCount } from '../../../utils/calculateApartmentCount'
 import MapProjectCard from './MapProjectCard';
 import MapProjectPopupCard from './MapProjectPopupCard';
 import useSessionStorageState from '../../../../../hooks/useSessionStorageState';
+import getInitialMapPosition from '../../../utils/getInitialMapPosition';
 
 import css from './MapResults.module.scss';
 
@@ -106,39 +107,6 @@ const MapResults = ({
     return icon;
   };
 
-  // Centers the map to middle of all the projects on initial render
-  const getInitialPosition = (): [number, number] => {
-    if (searchResults.length > 0) {
-      let maxLon: number = searchResults[0].coordinate_lon;
-      let minLon: number = searchResults[0].coordinate_lon;
-      let maxLat: number = searchResults[0].coordinate_lat;
-      let minLat: number = searchResults[0].coordinate_lat;
-
-      searchResults.forEach((x) => {
-        const { coordinate_lon, coordinate_lat } = x;
-        if (maxLon < coordinate_lon) {
-          maxLon = coordinate_lon;
-        }
-        if (!minLon || minLon > coordinate_lon) {
-          minLon = coordinate_lon;
-        }
-        if (!maxLat || maxLat < coordinate_lat) {
-          maxLat = coordinate_lat;
-        }
-        if (!maxLat || maxLat < coordinate_lat) {
-          minLat = coordinate_lat;
-        }
-      });
-
-      const lon = (maxLon - minLon) / 2 + minLon;
-      const lat = (maxLat - minLat) / 2 + minLat;
-
-      return [lat, lon];
-    }
-
-    return [60.17, 24.94];
-  };
-
   return (
     <section className={css.container} aria-label={header}>
       <header>
@@ -166,7 +134,7 @@ const MapResults = ({
       <div className={css.mapContainer}>
         <div className={css.mapWrapper} id={'asuReactMap'}>
           <MapContainer
-            center={getInitialPosition()}
+            center={getInitialMapPosition(searchResults)}
             zoom={12}
             maxBounds={[
               [59.4, 23.8],
